Move the Python wrapper code for view to be under the same directory
