+++ cpp/binary/viewer/CMakeLists.txt
@@ -38,8 +38,6 @@ qt_add_executable(
     ${MODMESH_PYTHON_SOURCES}
     ${MODMESH_PYMOD_MODMESH_HEADERS}
     ${MODMESH_PYMOD_MODMESH_SOURCES}
-    ${MODMESH_PYMOD_VIEW_HEADERS}
-    ${MODMESH_PYMOD_VIEW_SOURCES}
     ${MODMESH_PYMOD_SPACETIME_HEADERS}
     ${MODMESH_PYMOD_SPACETIME_SOURCES}
 )
