+++ cpp/modmesh/python/CMakeLists.txt
@@ -4,7 +4,6 @@
 cmake_minimum_required(VERSION 3.16)
 
 add_subdirectory(wrapper/modmesh)
-add_subdirectory(wrapper/view)
 add_subdirectory(wrapper/spacetime)
 
 set(MODMESH_PYTHON_HEADERS
