--- cpp/modmesh/python/wrapper/view/view.cpp
@@ -1,64 +1,0 @@
-/*
- * Copyright (c) 2022, Yung-Yu Chen <yyc@solvcon.net>
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * - Redistributions of source code must retain the above copyright notice,
- *   this list of conditions and the following disclaimer.
- * - Redistributions in binary form must reproduce the above copyright notice,
- *   this list of conditions and the following disclaimer in the documentation
- *   and/or other materials provided with the distribution.
- * - Neither the name of the copyright holder nor the names of its contributors
- *   may be used to endorse or promote products derived from this software
- *   without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-#include <modmesh/python/wrapper/view/view.hpp> // Must be the first include.
-#include <pybind11/stl.h>
-
-#include <modmesh/modmesh.hpp>
-#include <modmesh/python/common.hpp>
-
-namespace modmesh
-{
-
-namespace python
-{
-
-struct view_pymod_tag;
-
-template <>
-OneTimeInitializer<view_pymod_tag> & OneTimeInitializer<view_pymod_tag>::me()
-{
-    static OneTimeInitializer<view_pymod_tag> instance;
-    return instance;
-}
-
-void initialize_view(pybind11::module & mod)
-{
-    auto initialize_impl = [](pybind11::module & mod)
-    {
-        wrap_view(mod);
-    };
-
-    OneTimeInitializer<view_pymod_tag>::me()(mod, initialize_impl);
-}
-
-} /* end namespace python */
-
-} /* end namespace modmesh */
-
-// vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
