+++ cpp/modmesh/view/CMakeLists.txt
@@ -12,6 +12,7 @@ set(MODMESH_VIEW_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/RStaticMesh.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RMenu.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/view.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/wrap_view.hpp
     CACHE FILEPATH "" FORCE
 )
 
@@ -23,6 +24,7 @@ set(MODMESH_VIEW_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/RPythonText.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RStaticMesh.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/RMenu.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/wrap_view.cpp
     CACHE FILEPATH "" FORCE
 )
