+++ cpp/modmesh/view/view.hpp
@@ -37,9 +37,4 @@
 #include <modmesh/view/RStaticMesh.hpp>
 #include <modmesh/view/RAxisMark.hpp>
 
-namespace modmesh
-{
-
-} /* end namespace modmesh */
-
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
